Improve tests creating a mother test class
