+++ FrameworkSystem/test/testLoggerWrapper/tests/TestLogRecordCreation.py
@@ -3,12 +3,10 @@ Test LogRecord Creation
 """
 # imports
 import unittest
-import logging
-import sys
-from StringIO import StringIO
 
 # sut
 from DIRAC import gLogger, oldgLogger
+from DIRAC.FrameworkSystem.test.testLoggerWrapper.tests.TestLoggerWrapper import TestLoggerWrapper
 
 
 def cleaningLog(log):
@@ -20,28 +18,11 @@ def cleaningLog(log):
   return log
 
 
-class TestLogRecordCreation(unittest.TestCase):
+class TestLogRecordCreation(TestLoggerWrapper):
   """
   Test the creation of the different log records 
   via the always, notice, ..., fatal methods.
   """
-
-  def setUp(self):
-    """
-    Initialize at debug level with a sublogger and a special handler
-    """
-    gLogger.setLevel('debug')
-    self.log = gLogger.getSubLogger('log')
-    self.buffer = StringIO()
-
-    oldgLogger.setLevel('debug')
-    self.oldlog = oldgLogger.getSubLogger('log')
-    self.oldbuffer = StringIO()
-    sys.stdout = self.oldbuffer
-
-    # modify the output to capture the log into a buffer
-    if logging.getLogger().handlers:
-      logging.getLogger().handlers[0].stream = self.buffer
 
   def test_00always(self):
     """
